requirements.txt
streamlit
openai>=1.0
newspaper3k
beautifulsoup4
requests
python-dotenv

